doc: spike approach to ensure less renders happen on updating chain data
This still needs clean up but works in reducing the number of times that the code is called in generating the chainOrder value
